v3 minimal state
