+++ beacon-chain/state/v3/BUILD.bazel
@@ -6,6 +6,8 @@ load("//tools:ssz.bzl", "SSZ_DEPS", "ssz_gen_marshal")
 go_library(
     name = "go_default_library",
     srcs = [
+        "beacon_state.go",
+        "beacon_state_minimal.go", # keep
         "deprecated_getters.go",
         "deprecated_setters.go",
         "field_roots.go",
